# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "college management",
      "contact": {}
    },
    "host": "localhost:3000",
    "basePath": "/api/student",
    "securityDefinitions": {},
    "schemes": ["http"],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/forgetPassword": {
        "post": {
          "summary": "Forget Password",
          "tags": [
            "Student"
          ],
          "operationId": "ForgetPassword",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/ForgetPasswordRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
      "/verifyOtpCode": {
        "post": {
          "summary": "verify Otp Code",
          "tags": [
            "Student"
          ],
          "operationId": "verifyOtpCode",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/verifyOtpCodeRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
      "/changePassword": {
        "post": {
          "summary": "changePassword",
          "tags": [
            "Student"
          ],
          "operationId": "changePassword",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/changePasswordRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "ForgetPasswordRequest": {
        "title": "ForgetPasswordRequest",
        "example": {
          "email": "suhagiyadivyesh7600@gmail.com"
        },
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "verifyOtpCodeRequest": {
        "title": "verifyOtpCodeRequest",
        "example": {
          "email": "suhagiyadivyesh7600@gmail.com",
          "otpCode": "162813",
          "newPassword": "divyesh@new",
          "confirmPassword": "divyesh@new"
        },
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "otpCode": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "otpCode",
          "newPassword",
          "confirmPassword"
        ]
      },
      "changePasswordRequest": {
        "title": "changePasswordRequest",
        "example": {
          "email": "suhagiyadivyesh7600@gmail.com",
          "oldPassword": "divyesh@new",
          "newPassword": "divyesh@123",
          "confirmPassword": "divyesh@123"
        },
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "oldPassword",
          "newPassword",
          "confirmPassword"
        ]
      },
...
    },
    "security": [
      {
        "bearer": []
      }
    ],
...
  }